+++ R/tabula-deprecated.R
@@ -1,0 +1,6 @@
+#' Deprecated Functions in tabula
+#'
+#' These functions still work but will be removed (defunct) in the next version.
+#' @name tabula-deprecated
+#' @keywords internal
+NULL